12.07 JavaEE JSP 커밋 QnA 게시판 작성

+++ BoardApp/WebContent/imageboard/delete.jsp
@@ -1,3 +1,4 @@
+<%@page import="java.io.File"%>
 <%@page import="board.model.ImageBoard"%>
 <%@page import="board.model.ImageBoardDAO"%>
 <%@ page contentType="text/html;charset=utf-8"%>
@@ -7,17 +8,25 @@
 <%
 
 	String board_id = request.getParameter("board_id");
-
-	ImageBoardDAO dao = new ImageBoardDAO();
-	ImageBoard imageBoard = new ImageBoard();
+	String filename = request.getParameter("filename");
 	
-	imageBoard = dao.select(Integer.parseInt(board_id));
-	int result = dao.delete(imageBoard);
+	File file = new File("D:/IT_Korea_Class/JavaEE/BoardApp/WebContent/data/"+filename);
+	if(file.delete()) {
+		ImageBoardDAO dao = new ImageBoardDAO();
+		ImageBoard imageBoard = new ImageBoard();
+		
+		imageBoard = dao.select(Integer.parseInt(board_id));
+		int result = dao.delete(imageBoard);
+		
+		if(result==0){
+			out.print(getMsgBack("삭제 실패!"));
+		}else{
+			out.print(getMsgURL("삭제 성공!", "list.jsp"));
+		}	
+	} else {
+		out.print(getMsgBack("파일삭제 실패!"));
+	}
 	
-	if(result==0){
-		out.print(getMsgBack("삭제 실패!"));
-	}else{
-		out.print(getMsgURL("삭제 성공!", "list.jsp"));
-	}	
+	
 	
 %>
